refactor: move overlay pages to pages/overlay/; rename PauseScene to PauseOverlay
